Update SevenSegmentClockNrf_2021.ino
With this commit the first version of the ClockNet solution is finished.
We now hav a main node, listenig to an external sensor for temperature, humidity and UV index. This information is them distribbuted to a couple (4) ChildNodes.
Che ChildNodes has its own clock and requests the time from the controller but receives the outdoor information from the Main node, This means tha that the external sensor wich is battery powered only need to send information once to the main node. Each Child Node will receive the outdoor information as soon as the external sensor sends an update. ChilcNodes never requests data frpm T´the main node or the external sensor..