Correction for temperature precision
Corrected that the switch TEMPERATURE_PRECISION had no effect on the program.
If selected any of the precision alternative, default value od 12 bits always was set.
The correction did nt change any functionality for now, but if the precision is needed to change later to a lower precision the switches are working.